**What type of PR is this?**

/kind feature

**What this PR does / why we need it**:

Optimize spreading when using `kubernetes.io/hostname` as a topology key. We avoid calculating per-node counters during PreScore and do it in Score instead. This is faster because there are less map insertions and accesses.

```
default spreading

BenchmarkTestSelectorSpreadPriority/100nodes-56         	    5161	    226371 ns/op
BenchmarkTestSelectorSpreadPriority/1000nodes-56        	     788	   1516032 ns/op

master

BenchmarkTestDefaultEvenPodsSpreadPriority/100nodes-56         	    2168	    489570 ns/op
BenchmarkTestDefaultEvenPodsSpreadPriority/1000nodes-56        	     378	   3022404 ns/op

this PR

BenchmarkTestDefaultEvenPodsSpreadPriority/100nodes-56         	    3805	    329776 ns/op
BenchmarkTestDefaultEvenPodsSpreadPriority/1000nodes-56        	     631	   1858747 ns/op
```

This makes the implementation of default spreading with "topology spreading" closer to the default spreading:
- 0.3ms for 100 nodes, 1,000 pods (1.4x from default spreading)
- 1.8ms for 1,000 nodes, 10,000 pods (1.2x from default spreading)

**Special notes for your reviewer**:

The first commit contains the optimization.
The second commit updates the benchmark to reflect scheduler code, which parallelizes score calculation.

**Does this PR introduce a user-facing change?**:

```release-note
NONE
```