@alculquicondor 2020-03-30
/remove-sig scheduling